LPS-138503 account-admin-web: moves common logic to a base ScreenNavigationCategory
This is a preparatory step for adding a new define permissions
ScreenNavigationEntry.
